+++ src/TechDivision/ApplicationServer/Api/Node/ProvisionerNode.php
@@ -1,0 +1,68 @@
+<?php
+
+/**
+ * TechDivision\ApplicationServer\Api\Node\ProvisionerNode
+ *
+ * PHP version 5
+ *
+ * @category   Appserver
+ * @package    TechDivision_ApplicationServer
+ * @subpackage Api
+ * @author     Tim Wagner <tw@techdivision.com>
+ * @copyright  2013 TechDivision GmbH <info@techdivision.com>
+ * @license    http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link       http://www.appserver.io
+ */
+
+namespace TechDivision\ApplicationServer\Api\Node;
+
+/**
+ * DTO to transfer the provisioner information.
+ *
+ * @category   Appserver
+ * @package    TechDivision_ApplicationServer
+ * @subpackage Api
+ * @author     Tim Wagner <tw@techdivision.com>
+ * @copyright  2013 TechDivision GmbH <info@techdivision.com>
+ * @license    http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link       http://www.appserver.io
+ */
+class ProvisionerNode extends AbstractNode implements ProvisionerNodeInterface
+{
+
+    /**
+     * The provisioner name.
+     *
+     * @var string
+     * @AS\Mapping(nodeType="string")
+     */
+    protected $name;
+
+    /**
+     * The provisioner type.
+     *
+     * @var string
+     * @AS\Mapping(nodeType="string")
+     */
+    protected $type;
+
+    /**
+     * Returns the provisioner type.
+     *
+     * @return string The provisioner type
+     */
+    public function getType()
+    {
+        return $this->type;
+    }
+
+    /**
+     * Returns the provisioner name.
+     *
+     * @return string The provisioner name
+     */
+    public function getName()
+    {
+        return $this->name;
+    }
+}