+++ src/TechDivision/ApplicationServer/StandardProvisioner.php
@@ -7,8 +7,8 @@
  *
  * @category  Appserver
  * @package   TechDivision_ApplicationServer
- * @author    Johann Zelger <j.zelger@techdivision.com>
- * @copyright 2013 TechDivision GmbH <info@techdivision.com>
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2014 TechDivision GmbH <info@techdivision.com>
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  * @link      http://www.appserver.io
  */
@@ -16,19 +16,18 @@
 namespace TechDivision\ApplicationServer;
 
 use TechDivision\ApplicationServer\Api\Node\ProvisionNode;
-use TechDivision\ApplicationServer\Interfaces\ProvisionerInterface;
 
 /**
  * Standard provisioning functionality.
  *
  * @category  Appserver
  * @package   TechDivision_ApplicationServer
- * @author    Johann Zelger <j.zelger@techdivision.com>
- * @copyright 2013 TechDivision GmbH <info@techdivision.com>
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2014 TechDivision GmbH <info@techdivision.com>
  * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
  * @link      http://www.appserver.io
  */
-class StandardProvisioner implements ProvisionerInterface
+class StandardProvisioner extends AbstractProvisioner
 {
 
     /**
@@ -44,34 +43,6 @@ class StandardProvisioner implements ProvisionerInterface
      * @var string
      */
     const DEFAULT_CONFIGURATION = '/etc/appserver.d/provision.xml';
-
-    /**
-     * The containers base directory.
-     *
-     * @var string
-     */
-    protected $service;
-
-    /**
-     * The initial context instance.
-     *
-     * @var \TechDivision\ApplicationServer\InitialContext
-     */
-    protected $initialContext;
-
-    /**
-     * Contructor
-     *
-     * @param \TechDivision\ApplicationServer\InitialContext $initialContext The initial context instance
-     */
-    public function __construct($initialContext)
-    {
-        // add initialContext
-        $this->initialContext = $initialContext;
-
-        // init API service to use
-        $this->service = $this->newService('TechDivision\ApplicationServer\Api\ProvisioningService');
-    }
 
     /**
      * Provisions all web applications.
@@ -142,9 +113,6 @@ class StandardProvisioner implements ProvisionerInterface
     protected function executeProvision(ProvisionNode $provisionNode, \SplFileInfo $webappPath)
     {
 
-        // load the service instance
-        $service = $this->getService();
-
         // load the steps from the configuration
         $stepNodes = $provisionNode->getInstallation()->getSteps();
 
@@ -180,16 +148,6 @@ class StandardProvisioner implements ProvisionerInterface
     }
 
     /**
-     * Returns the servers deploy directory
-     *
-     * @return string
-     */
-    public function getWebappsDir()
-    {
-        return $this->getService()->getWebappsDir();
-    }
-
-    /**
      * Returns the absolute path to the appservers PHP executable.
      *
      * @return string The absolute path to the appserver PHP executable
@@ -197,38 +155,5 @@ class StandardProvisioner implements ProvisionerInterface
     public function getAbsolutPathToPhpExecutable()
     {
         return $this->getService()->realpath(StandardProvisioner::PHP_EXECUTABLE);
-    }
-
-    /**
-     * (non-PHPdoc)
-     *
-     * @param string $className The API service class name to return the instance for
-     *
-     * @return \TechDivision\ApplicationServer\Api\ServiceInterface The service instance
-     * @see \TechDivision\ApplicationServer\InitialContext::newService()
-     */
-    public function newService($className)
-    {
-        return $this->getInitialContext()->newService($className);
-    }
-
-    /**
-     * Returns the inital context instance.
-     *
-     * @return \TechDivision\ApplicationServer\InitialContext The initial context instance
-     */
-    public function getInitialContext()
-    {
-        return $this->initialContext;
-    }
-
-    /**
-     * Returns the service instance to use.
-     *
-     * @return \TechDivision\ApplicationServer\Api\ServiceInterface $service The service to use
-     */
-    public function getService()
-    {
-        return $this->service;
     }
 }